+++ app/assets/javascripts/student_profile/page_container.jsx
@@ -3,12 +3,12 @@ import MixpanelUtils from '../helpers/mixpanel_utils.jsx';
 import PropTypes from '../helpers/prop_types.jsx';
 import {fromPair} from '../helpers/from_pair.jsx';
 import {merge} from '../helpers/react_helpers.jsx';
+import Api from '../student_profile/api.js'
 
 (function() {
   window.shared || (window.shared = {});
   const Routes = window.shared.Routes;
   const StudentProfilePage = window.shared.StudentProfilePage;
-  const Api = window.shared.Api;
   /*
   Holds page state, makes API calls to manipulate it.
   */
